refactor: migrar stores zustand por dominio com adaptadores
