+++ src/app/routes/dashboard/pages/pos.tsx
@@ -15,9 +15,9 @@ import {
 } from "lucide-react"
 import { useEffect, useMemo, useRef, useState } from "react"
 import { toast } from "sonner"
-import { useCartStore } from "@/app/store/cart-store"
-import { useRestaurantStore } from "@/app/store/restaurant-store"
 import { Modal } from "@/components/ui/modal"
+import { useCartStore } from "@/domains/pos/store/cart-store"
+import { useRestaurantStore } from "@/domains/restaurant/store/restaurant-store"
 import { useCategories } from "@/hooks/use-categories"
 import { useOrganizationCheck } from "@/hooks/use-organization-check"
 import { useProducts } from "@/hooks/use-products"
